+++ cmd/curator/main.go
@@ -22,7 +22,7 @@ import (
 )
 
 const (
-	version = "0.22.2"
+	version = "0.22.3"
 )
 
 func main() {